move demo -> ttopt-demo
