+++ MANIFEST.in
@@ -1,4 +1,4 @@
-recursive-include demo *
+recursive-include ttopt-demo *
 include LICENSE.txt
 include README.md
 include requirements.txt
