+++ ttopt-demo/vect.py
@@ -1,21 +1,21 @@
-"""The demo of using ttopt. Example with QTT for 100-dimensional function.
+"""The demo of using ttopt. Simple example with "scalar" input.
 
-We'll find the minimum for the 100-dimensional Alpine function with vectorized
-input. The target function for minimization has the form f(X), where input X is
-the [samples, dimension] numpy array.
+We'll find the minimum for the simple analytic function of the form f(x), where
+input x is the [dimension] numpy array (not vectorized!).
 
-Run it from the root of the project as "python demo/qtt_100d.py".
+Run it from the root of the project as "python ttopt-demo/vect.py".
 
 As a result of the script work we expect the output in console like this:
 "
 ...
-Alpine-100d | evals=1.00e+05 | t_cur=2.87e-01 | e_x=1.16e+01 e_y=2.29e-02
+Simple-5d  | evals=1.00e+04 | t_cur=1.38e-01 | e_x=3.55e-02 e_y=1.03e-04
 ----------------------------------------------------------------------
-Alpine-100d | evals=1.00e+05 | t_all=6.89e+00 | e_x=1.16e+01 e_y=2.29e-02
+Simple-5d  | evals=1.00e+04 | t_all=1.53e-01 | e_x=3.55e-02 e_y=1.03e-04 
 "
 
 """
 import numpy as np
+from scipy.optimize import rosen
 
 
 from ttopt import TTOpt
@@ -25,24 +25,24 @@ from ttopt import ttopt_init
 np.random.seed(42)
 
 
-d = 100                     # Number of function dimensions:
+d = 5                       # Number of function dimensions:
 rank = 4                    # Maximum TT-rank while cross-like iterations
-def f(X):                   # Target function
-    return np.sum(np.abs(X * np.sin(X) + 0.1 * X), axis=1)
+def f(x):                   # Target function
+    return np.sin(0.1 * x[0])**2 + 0.1 * np.sum(x[1:]**2)
 
 
 # We initialize the TTOpt class instance with the correct parameters:
 tto = TTOpt(
     f=f,                    # Function for minimization. X is [samples, dim]
     d=d,                    # Number of function dimensions
-    a=-10.,                 # Grid lower bound (number or list of len d)
-    b=+10.,                 # Grid upper bound (number or list of len d)
-    p=2,                    # The grid size factor (there will n=p^q points)
-    q=12,                   # The grid size factor (there will n=p^q points)
-    evals=1.E+5,            # Number of function evaluations
-    name='Alpine',          # Function name for log (this is optional)
-    x_opt_real=np.ones(d),  # Real value of x-minima (x; this is for test)
+    a=-1.,                  # Grid lower bound (number or list of len d)
+    b=+1.,                  # Grid upper bound (number or list of len d)
+    n=2**6,                 # Number of grid points (number or list of len d)
+    evals=1.E+4,            # Number of function evaluations
+    name='Simple',          # Function name for log (this is optional)
+    x_opt_real=np.zeros(d), # Real value of x-minima (x; this is for test)
     y_opt_real=0.,          # Real value of y-minima (y=f(x); this is for test)
+    is_vect=False,          # The function accepts only one spatial point
     with_log=True)
